requirements.txt
opencv-python
numpy
python-chess
pygame
cairosvg